refactor: 카 이름 클래스 모듈화
