+++ src/main/java/racingcar/domain/RacingCar.java
@@ -5,10 +5,9 @@ import racingcar.domain.strategy.MoveStrategy;
 
 public class RacingCar implements Car {
 
-    private static final int NAME_SIZE = 5;
     private static final int START_POSITION = 0;
 
-    private final String name;
+    private final Name name;
     private int position;
 
     public RacingCar(String name) {
@@ -16,8 +15,7 @@ public class RacingCar implements Car {
     }
 
     public RacingCar(String name, int position) {
-        isCorrectNameSize(name);
-        this.name = name;
+        this.name = new Name(name);
         this.position = position;
     }
 
@@ -32,17 +30,8 @@ public class RacingCar implements Car {
         return this.position == position;
     }
 
-    private void isCorrectNameSize(String name) {
-        if (name == null || name.isEmpty()) {
-            throw new IllegalArgumentException("자동차 이름이 빈값입니다!");
-        }
-        if (NAME_SIZE < name.length()) {
-            throw new IllegalArgumentException("자동차 이름의 길이가 5를 초과했습니다!");
-        }
-    }
-
     public String getName() {
-        return name;
+        return name.getName();
     }
 
     public int getPosition() {
